+++ ENGIN/ENGIN_DEMO/source/GameObjectTestScene.cpp
@@ -15,7 +15,7 @@ GameObjectTestScene::GameObjectTestScene(Engin::Engin* engine) : gameObject("mai
 	this->engine = engine;
 	alpha = 0.0f;
 
-	camera->initCamera(0.0f, 0.0f, engine->getWindow().getWindowWidth(), engine->getWindow().getWindowHeight(), 0.0f, 0.0f, engine->getWindow().getWindowWidth() / 2, engine->getWindow().getWindowHeight()/2);
+	camera->initCamera(0.0f, 0.0f, static_cast<GLfloat>(engine->getWindow().getWindowWidth()), static_cast<GLfloat>(engine->getWindow().getWindowHeight()), 0.0f, 0.0f, engine->getWindow().getWindowWidth() / 2.0f, engine->getWindow().getWindowHeight() / 2.0f);
 
 	camera->setZoomLevel(0.5f);
